adding extended feels review options
sketching out a submenu format, and testing it by making a new feels
reviewing option. if the user is publishing, review feels now includes
the option to modify publishing settings for individual entries (which
is a function yet to be written.)

maybe i'll remember to document this to make a guide for adding new
submenus.